apps/kbd-ir/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pydantic==1.10.15
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.122
python-multipart==0.0.9
pytest==8.2.2
